Issue long in admin api

See merge request c1718515/swiperightbrains!33

+++ src/main/java/com/nsa/cubric/application/repositories/AccountRepositoryStatic.java
@@ -18,6 +18,6 @@ public interface AccountRepositoryStatic {
 
     ProfileDTO getProfileByEmail(String email);
 
-    public boolean removeUser(Integer userId);
-    public Long removeUserResponses(Integer userId);
+    public boolean removeUser(Long userId);
+    public Long removeUserResponses(Long userId);
 }
